Separate extensions into new files
